Enable helm/armada plugin delivery with the application
This creates a new package spec called python-k8sapp-platform that will
build two packages: python-k8sapp-platform and
python-k8sapp-platform-wheels. These packages will hold all the
stevedore plugins needed to support the application.

These packages are included in the build dependencies for the
stx-platform-helm application package build where the wheels file will
be included in the application tarball.

The helm and armada plugins have been relocated to this repo and
provided in a k8sapp_platform python module. This module will be
extracted from the wheels and installed on the platform via the sysinv
application framework. The module will be made available when the
application is enabled.

Now that the plugins and the application are residing in the same
repository, GITREVCOUNT will be used to drive the package version of
both to keep them in sync.

Tox test support has been provided for the plugin module.

Change-Id: I2085440cd23790ef49662b427d2e894b515b6f2a
Depends-On: https://review.opendev.org/#/c/686971/
Story: 2006537
Task: 36931
Signed-off-by: Robert Church <robert.church@windriver.com>