Fixed the default storageclass and replicas number
A previous update that added a new storageclass
using cephfs-provisioner affects the installation of
some application that are trying to use the
default storageclass.

When an application will try to use the default
storageclass it will find two default storageclasses
and we want to avoid this problem by using a single
default storageclass that can be changed by the user
through helm overrides.

Also this fixes the replicas number for the deployment,
it was hardcoded with 1, but now it can be changed by the user.

This commit address an application issue with
https://review.opendev.org/c/starlingx/platform-armada-app/+/758786

Signed-off-by: Daniel Safta <daniel.safta@windriver.com>
Story: 2008162
Task: 40908
Change-Id: Ib1e4db29397e674a17d45120f6c0b3babaf1880b