Removed extra serviceAccount from cephfs-provisioner
cephfs-provisioner may need to create new resources
in the kubernetes cluster. It was granted access to
some of the resources including namespaces but when
https://review.opendev.org/c/starlingx/platform-armada-app/+/778746
got merged the serviceAccount was changed.

I have updated the serviceAccount with access to
creating new namespaces and secrets.

The serviceAccount that was initially used to create
namespaces and secrets is not needed anymore, so I
have removed it.

Closes-bug: 1921197
Change-Id: I3c683776f3ecaf9c78d1a6b5b1108e9582497dde
Signed-off-by: Daniel Safta <daniel.safta@windriver.com>