Fix chart syntax to comply with Go 1.14
When testing patch [1] it was found that the usage of parenthesis on
this chart doesn't comply with the Tiller implementation on Go 1.14.
This fix will allow the upversion of Tiller to > 2.16.8 which is
required for the Armada upversioning.

[1] https://review.opendev.org/c/starlingx/ansible-playbooks/+/792297

Partial-Bug: #1928994
Signed-off-by: Thiago Brito <thiago.brito@windriver.com>
Change-Id: I8532036cd70b3e20a143dceba19119779413eb39
(cherry picked from commit 7d30242557401bfdb0be5f75bd29ce05708d250f)