Add a toleration in ceph-pools-audit
CMK will check worker has the taint "cmk=true:NoSchedule" after assign
application-isolated cores on Applications. Current ceph-pools-audit
Cron job does not allow pod to be scheduled when this taint is applied.
It is observed one ceph-pools-audit pod is pending because it can not
tolerate {cmk:true} taint on AIO-SX. Add toleration support for
ceph-pools-audit so that a helm override can be provided to allow job
scheduling when CMK is installed.

Closes-Bug: 1934006
Signed-off-by: Yvonne Ding <yvonne.ding@windriver.com>
Change-Id: I57462ac5d4ce700dbd89dd7569101834e0f0c83d