Update tox.ini to work with tox 4
This change will allow this repo to pass zuul now
that this has merged:
https://review.opendev.org/c/zuul/zuul-jobs/+/866943

Tox 4 deprecated whitelist_externals.
Replace whitelist_externals with allowlist_externals

Zuul is pinned to use tox < 4 in order for the py39
unit tests to pass. Additional changes for skipsdist
may be required in order to work properly with tox 4.

Partial-Bug: #2000399

Signed-off-by: Al Bailey <al.bailey@windriver.com>
Change-Id: Iab988c50b18d683a0d0967cbcc31bbad621005f8