Adjust platform docker images to get from registry.local
The platform-integ-apps docker images were pointing to the
public registry. This was caused because the path for each
image was defined in values.yml file for cephfs and rbd provisioners,
but not defined in the static overrides files.

This change adds the image paths to the static overrides files, so
Conductor can change the prefix to registry.local:9001 in the
deployments.

Test-Plan:
  PASS: Fresh install AIO-SX and apply platform-integ-apps
        application. Verify if the deployments have all images
        pointing to registry.local:9001 registry.

Closes-bug: 2015788

Change-Id: Icb04d4fb9c9073eeb8de4bd9507a8460f0fbda14
Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>