Remove debug option from cephfs PVs
Since version v3.9.0 of ceph-csi, the "debug" option has been
removed from mountOptions of the cephfs storage class, however,
this option still exists on cephfs PVs created with the previous
version of ceph-csi, causing the pod to fail.

To resolve this, a check for existing cephfs PVs has been added
to the cephfs storage-init script to remove this parameter if
it exists.

Test Plan:
  PASS: Create a PVC and pod on AIO-SX with ceph-csi v3.6.2
  PASS: Build platform-integ-apps with changes (ceph-csi v3.9.0)
  PASS: Check that "mountOption: -debug" is not present in the
        cephfs storage class and pv.

Closes-Bug: 2047571

Change-Id: Id7c8f77d2bc0b4e4afc67966810d5d3c40fc1e06
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>