Separates the type of operation in the semantic check condition
The change made in [1] created a bug when creating a new storage
using a new tier, where the storageclass was not created.

This happened because when it is an operation of type
APP_EVALUATE_REAPPLY_OP, relative_timing is not sent in lifecycle,
and this caused an exception that was caught by the conductor:

ERROR sysinv.conductor.manager [-] Unexpected error during hook
for app platform-integ-apps, error: Cannot load 'relative_timing'
in the base class: NotImplementedError: Cannot load 'relative_timing'
in the base class.

To resolve this, the condition to check relative_timing only if the
operation is APP_APPLY_OP was corrected.

[1]: https://review.opendev.org/c/starlingx/platform-armada-app/+/907246

Test Plan:
  PASS: Apply pkg with this changes
  PASS: Create new storage using new tier

Closes-bug: 2054738

Change-Id: I6ff7180e339a55ac5dfe20e378cb301e52eaa6a3
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>