Remove unnecessary code for stx.10.0
This commit removes code from the lifecycle script that was only
needed for stx.8.0 -> stx.9.0 upgrades.

The purpose of the code was to update the user overrides of
'classes' (when existing) to 'storageClasses', a change
that occurred in the version of platform-integ-apps
from stx.8.0 to stx.9.0.

Since the platform upgrade to stx.10.0 will only be allowed from
stx.9.0, and the platform-integ-apps of both versions will have
'storageClasses', the code is no longer necessary.

In addition to no longer needing the code, it caused an issue.
In a scenario where there were user overrides of 'classes'
before the app update, every time there is an apply with new
overrides after the update, the value to be considered
returned to the 'classes' overrides of the old app version
just with the name change. In other words, after the update in
this scenario, it is not possible to use new overrides for
'storageClasses'. Removing the code fixes this.

Closes-bug: 2076461

Test Plan:
 PASS: Build a new app package with the code changes
 PASS: Successfully execute upload, apply, remove, update and delete
       operations for 'platform-integ-apps' on both AIO-SX and
       AIO-DX environments.
 PASS: Using an old app version, create user overrides for 'classes'.
       Then, update the application to the master version (without
       the current changes, this update will change 'classes' to
       'storageClasses'). Next, update the app to the version built
       with the changes from this review and test the use of new
       user overrides for 'storageClasses'.

Change-Id: Ia3d8cd2bdeb3a60bf97e005e56899d1a0965042c
Signed-off-by: Gabriel de Araújo Cabral <gabriel.cabral@windriver.com>