Update ceph-csi helm charts to v3.15.0
Update RBD/CephFS provisioners to v3.15.0 based on
https://github.com/ceph/ceph-csi/releases/tag/v3.15.0

We use the latest versions of cephcsi, csi-provisioner,
csi-attacher, csi-snapshotter, csi-node-driver-registrar,
csi-resizer and ceph-config-helper:

- cephcsi v3.15.0
- csi-provisioner v5.3.0
- csi-attacher v4.10.0
- csi-snapshotter v8.3.0
- csi-node-driver-registrar v2.15.0
- csi-resizer v1.14.0
- ceph-config-helper ubuntu_jammy_18.2.2-1-20241210

Two patches were replaced (14 and 15) which were about creating
the snapshotclass.yaml files for both RBD and CephFs. This is not
needed anymore since in cephcsi v3.15.0 these files already exist
and instead of snapshotClass, the main field is called
volumeSnapshotClass. We altered the existing snapshotclass.yaml
files and values.yaml to use the old names and ensure backwards
compatibility.

This are the field names changed in 3.15.0 upstream release we
reverted back:

- snapshotClass	-> volumeSnapshotClass
- provisionerSecretNamespace -> snapshotterSecretNamespace
- provisionerSecret -> snapshotterSecret

In v3.15.0 the csi-attaccher was added to cephfs
changing the attachRequired field from false to true
in the CSI driver definitions. Since this field is immutable,
we need to delete the CSI driver for cephfs before upgrading
or downgrading it, hence the change in the lifecycle.

Test Plan:
 PASS: App upload/apply/remove/delete/update/rollback
 PASS: Create PVC using storageclass general (rbd)
 PASS: Create pod using RBD PVC
 PASS: Resize RBD PVC
 PASS: Create RBD volume snapshot
 PASS: Create PVC using storageclass cephfs
 PASS: Create pod using cephfs PVC
 PASS: Resize cephfs PVC
 PASS: Create cephfs volume snapshot
 PASS: Enable ceph-pools-audit
 PASS: Change pool size and in the backend and check the pools
 PASS: B&R with and without the flag wipe_ceph_osds=true

* Tests were performed on SX using IPv4 with K8s 1.32.2-1.33.0

Change-Id: I9fc045645ccc7a1683b57a31a954dd7f044011d0
Signed-off-by: Heitor Farhat <heitor.farhat@windriver.com>