Force reapply of platform-integ-apps after deploy rollback
When we have ceph backend configured (platform-integ-apps applied)
with some pod using it, and a platform upgrade is done and then a
rollback, the platform-integ-apps is not configured correctly,
because when it is updated (downgraded) after the 'software deploy
activate-rollback', the platform version is still STX.11, so the
configured monitor IP is wrong, because in STX.11 the floating
address is used [1].

However, after 'software deploy host-rollback' and unlocking, the
conductor manager does not identify any changes to the overrides
in /opt/platform/helm, precisely because no files in this
directory were changed, as the modified overrides were in STX.11.

To solve this issue, the platform-integ-apps lifecycle receives a
downgrade operation, checks if it is a rollback, and then the N-1
version (stx.10) overrides have their contents erased, to ensure
there is a difference between the hash calculations and that the
app's reapply flag is set inside the 'evaluate_apps_reapply'
function.

[1]: https://review.opendev.org/c/starlingx/config/+/954367

Test Plan:
 - PASS: On AIO-SX, upgrade the platform and then rollback.
 - PASS: Check if all pods are running.

Closes-Bug: 2122354

Change-Id: I881a017763d9ac757af86b16c4aa4ad2863d3492
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>