Fix the ceph-pools-audit-helm version for compatibility with kube_app
In the platform-integ-apps upversion made in [1], the
ceph-pools-audit chart version was changed from 1.0.0 (1.0-0)
to 1.0.1 (1.0-1), however, in kube_app, the last value, which
refers to the patch, is disregarded and the auto-increment
strategy is applied, which uses the number of commits from
BASE_SRCREV, in addition to the value defined in 'stx_patch'.

Therefore, since the previous version was 1.0-0 and, using the
automatic increment strategy, it became 1.0-1, when generating
the chart after the upversion, it also shows the same version.

To fix this problem, it is necessary to modify the MINOR.
Therefore, the chart version was changed from 1.0-1 to 1.1-0.

[1]: https://review.opendev.org/c/starlingx/platform-armada-app/+/959860

Test Plan:
  PASS: Build app package and check the chart version.
  PASS: Update the applied app with this new package
	(system application-update).

Closes-Bug: 2131909

Change-Id: I59f83f1176e005fbbd199b5a271a8f94359905ec
Signed-off-by: Erickson Silva de Oliveira <Erickson.SilvadeOliveira@windriver.com>